pyproject.toml
[tool.poetry]
name = "etl-assainissement"
version = "1.0.0"
description = "Pipeline ETL pour assainissement de données financières et ventes"
authors = ["Charles Nguessan <charles@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
apache-airflow = ">=2.10.0,<3.0.0"
apache-airflow-providers-postgres = "^5.10.0"
apache-airflow-providers-mysql = "^5.5.0"
psycopg2-binary = "^2.9.11"
pymysql = "^1.1.2"
pandas = "^2.3.3"
python-dotenv = "^1.2.1"
sqlalchemy = ">=1.4.36,<2.0"
pyyaml = "^6.0.1"
pycountry = "^24.1.0"
numpy = "^1.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.12.1"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["."]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html"
]
markers = [
    "unit: Tests unitaires",
    "integration: Tests d'intégration",
    "slow: Tests lents"
]